Break down event "duck"
