+++ src/features/TopMenu/UserEventsMenu/UserEventsMenu.tsx
@@ -3,7 +3,7 @@ import { connect, MapDispatchToProps } from 'react-redux';
 import ListItem from 'src/components/core/ListItem';
 import Menu from 'src/components/core/Menu';
 import { StyleRulesCallback, withStyles, WithStyles } from 'src/components/core/styles';
-import { async } from 'src/store/events';
+import { markAllSeen } from 'src/store/events/event.request';
 import { MapState, ThunkDispatch } from 'src/store/types';
 import UserEventsButton from './UserEventsButton';
 import UserEventsList from './UserEventsList';
@@ -104,7 +104,7 @@ interface DispatchProps {
 
 const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (dispatch: ThunkDispatch) => ({
   actions: {
-    markAllSeen: () => dispatch(async.markAllSeen()),
+    markAllSeen: () => dispatch(markAllSeen()),
   },
 });
