+++ src/store/selectors/inProgressEvents.ts
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { isInProgressEvent } from 'src/store/events';
+import { isInProgressEvent } from 'src/store/events/event.helpers';
 
 export default (type: string) => createSelector<ApplicationState, Linode.Event[], Linode.Event[]>(
   (state) => state.events.events,
